Change CATCH to return NULL if no THROW is encountered
This adapts the CATCH function so that it doesn't pass through a non
thrown value any longer:

    >> catch [1 + 2]
    // null

Previously that would have been 3.  The new idea is that it is null,
making it easier to tell if a catch occurred.  It's easy enough to
have the old effect of a value "dropping out" by just having it
do a THROW of the final value.

This removes the need for a /WITH handler function being needed, as
one can just use THEN.  However, there was a bit of an issue with
named throws, as a handler would like to get both the name and the
value thrown.  This goes ahead and does that with a BLOCK!, whose
first element is the throw name and if there is a second value, it is
the thrown value (if the array is only length 1, then that means the
value thrown was a null).