Improve CSCAPE indenting behavior, work around REWORD
Previously, CSCAPE was able to take a block-like pattern and apply
line termination behavior and indentation for you.  This let you say:

   enum Some_Enum {
       $(Values),
   };

And it would realize to indent each line by that amount, and it would
put the comma on every line but the last because you used $().  (If
you had used $[], it would put the comma on the last line as well.)

This improves the generalization, so that not just indenting can be
preserved but any pattern leading the line gets repeated.  It also
corrects an issue where substitutions which happened to use the same
substitution but appeared in different contexts (e.g. multiple
appearances of value which had something other than a comma) were
messing up.  So this numbers the substitutions as it goes along so
as not to get confused.

There were some bugs in REWORD that were exposed during this process,
and rather than work around them this just changes to use REPLACE.
The REWORD issues are filed in the rebol-issues repository for later
reviewing.