Move extension init and shutdown into natives (incomplete)
Extensions previously had an Init() and Quit() function, which had to
have a prototype defined as well as be called by C code specifically.
This would be for initialization any OS services (such as cryptography
or sockets) that would have to be done in advance of running any
natives.

The existence of these functions as C hooks meant they could only be
called by C code...leading to questions of when and why they should
happen, plus what parameterization they should get.  It's not clear
what makes an extension special in wanting to run shutdown code just
because it happens to have some native code accompanying a module--a
completely user-written module might wish to do some cleanup as well.

So this moves that initialization such that it is done explicitly by
calling an initialization native when necessary as part of the
ordinary module code.  It's incomplete as there is no shutdown mechanic
for modules, but that will be added as a more general feature.

Due to this change, the "extension init" C file which every extension
was forced to have is no longer necessary, which streamlines the
extension creation process a bit.