Fix Module Isolation Problem
The pattern:

    not any [...] then [...]

Does not work as one might think, since NOT produces a LOGIC!.  It
can thus only be used with AND/OR in this way.  Alternately, you can
use NONE which is like NOT ANY but returns null:

     none [...] then [...]

This corrects one such mistake that had the effect of breaking the
module `Options: [isolate]` feature.