Run extension module initialization through LOAD-MODULE
Previously loading an extension went through a usermode helper called
LOAD-EXTENSION-HELPER which was a distinct code path from the general
mechanics which happen when a module is loaded.  So it was repeating
issues of binding and execution, and lacked support for options like
[isolate].

In order to prevent changes to a critical routine like PRINT from
affecting the console--as well as when any libRebol API functions might
write stray words or redefine things for a particular extension's
purposes, an isolation-type feature is necessary.  The seemingly best
path forward to doing that properly is to run it through the same
code that implements module isolation.

Hence this changes LOAD-MODULE to be able to take in a pre-made module
with some natives loaded into it, instead of creating its own.  The
natives have to reference the module as part of their definition,
because if they are to be run they need to be able to tell from the
stack what context to bind libRebol API calls in.  (e.g. if you say
rebRun("append") inside one module, it might not mean the same thing
as if you say it inside another one).

The chicken-and-egg problem of needing to call those natives during the
startup in modules is solved for the moment by loading modules "/INTO"
a pre-made module, which LOAD-EXTENSION creates.

While updating the headers to use `Type: Module`, `Options: [isolate]`
this also standardizes the headers somewhat, to use proper casing for
the option names...and standardizes references to the Apache 2.0
license and current copyright to Rebol Open Source contributors.