Move TCC to extension, un-fork and use latest TCC
This migrates the TCC compiler interface from being part of the core
executable (so that the stub natives aren't included in things like the
build for JavaScript).  That is one aspect of an important separation
of concerns.

But an even more important issue was that the initial implementation
method for the TCC extension involved using a forked and patched
version of TCC.  This is clearly less preferable than finding a way to
use TCC unmodified, for many reasons:

* Maintenance of the fork (as TCC would likely not take Rebol-specific
  changes to their master)

* If one didn't maintain the fork, then not getting updates to libtcc
  if it added newer/better features or platforms, or kept sync with
  chnanges on existing platform versions.

* Not being able to use a pre-built or shared library installation of
  libtcc on a system (e.g. gotten by `sudo apt-get libtcc`)

Also, the scenario for the TCC extension had focused fairly narrowly
on simple demos of "user natives" as a proof of concept.  It ran a demo
that calculated Fibonnacci numbers, which only requires the C language
and no includes besides Rebol APIs.  Real examples would be calling
many more C functions from libraries provided on the host, and also
an interesting scenario is to build targets to disk--e.g. to bootstrap
Rebol itself (one of the hopeful applications of TCC integration).

So the lion's share of the work of this commit is not merely moving
code out of the Rebol core to an extension, but all the concerns of
bringing TCC up to date (there have been changes in the intervening
years since Shixin wrote the original), and finding a way to simplify
the overall process to a less invasive and future-friendly way of
linking TCC functionality in via an extension.

One simplifying aspect is that the libRebol API has matured to the
point where %rebol.h should give all the functionality a TCC-built user
native would need.  Hence the dependency on %sys-core.h is removed.
Not having to build sys-core.h and link to it means that Rebol can
focus on exposing only the libRebol entry points, and that rebol.h
can be distributed as a single file.

Another is to move as much "front end" code for the COMPILE command
into usermode as possible.  Then a lower level COMPILE* command is
exposed as a native, which does the actual interfacing to TCC.

These are good first steps and can successfully run the Fibonnaci
example on Travis for Linux 32-bit and 64-bit.  However, there still
is work to do for giving a more "turnkey" experience of shipping with
necessary files so that the code will run places other than Travis,
and to have the libRebol symbols added using tcc_add_sym() by the
extension (this is necessary on Windows, but generally not on Linux).