Reorder DELIMIT parameters
The frequent use case of DELIMIT has a very simple delimiter string,
with a more complex expression producing the things to delimit (such
as a MAP-EACH).  While Rebol has a pattern of mutating operations
typically taking the series being operated on as the first parameter,
DELIMIT is generating a new series...and it's been fairly clear that
the source is more legible when the delimiter is first.

This goes ahead and makes the change.