Improve rebmake return value type annotations
The initial version of rebmake that was committed was theoretically
going to be able to target R3-Alpha, so it couldn't use features like
RETURN: type annotations.  Yet it was eventually decided that only
Ren-C builds would be used.  This goes ahead and marks up return types
on rebmake methods, to help with understanding what the code is doing.

Another feature that came about was the METHOD, which allowed picking
up member variables inside of something that also gathered locals.
While the actual implementation of method to "work" requires a newer
version of Ren-C than the one used as an r3-make currently, there
is an emulation that does well enough for the purposes of rebmake at
the moment.

This also brings in a simulated void to the shim layer, by using an
"ornery" function value that can errors if it is executed.  That
specific function can be recognized by VOID?, and other tricks are
used...including making a 0-arity RETURN in "procedure-like" actions.

Also disables the FFI build test, as part of the ongoing attempt to
get Travis to pass while the build system is being worked on.