Remove GROUP! rhs, non-BLOCK! lhs protection from UNLESS
When the XOR operator was introduced, its similarity to AND and OR
suggest that it might be called the same way in a non-short-circuit
fashion as `foo xor [bar]`.  However, there is no way to implement
XOR as short-circuit...it has to evaluate both sides.  This could
result in confusion that the block would be treated literally, so
only a GROUP! was accepted on the right (thus matching the non-short
circuit behavior of AND and OR, which evaluate the right hand side
even if they don't need to if a GROUP! is used).

When UNLESS was moved in to behave more like that family of functions,
it also was changed to require the GROUP! (as well as to have the
protection of AND, OR, XOR not to allow literal GROUP! on the left,
to prevent mistakes like **[x] or [y]**, where the need for the
BLOCK! on the right was mistaken for a need for a block on the left,
thus never running the right hand side).  However, while conceptually
in the same family, UNLESS looks very different and the protections
create unnecessary baggage.  It also disallows:

    data: [] unless ...

Which might be interesting if someone wants a BLOCK! (though that
block would be read-only in current designs).

This removes the protections, making UNLESS usable without GROUP!s in
more cases.