Add COLLECT-TEXT, COLLECT-LINES, NEWLINED
This introduces some convenience functions that resemble collect.

COLLECT-TEXT allows you to collect a single line of text whose KEEPs
of a block assumed are assumed to be unspaced portions, while the
individual KEEPs are then merged with spacing.

COLLECT-LINES allows you to collect a block of text lines, where each
KEEP is assumed to space any block it is given with SPACED.

NEWLINED is a slight modification on DELIMIT NEWLINE which also adds a
newline on the end.  DELIMIT usually doesn't put a delimiter on the
end because you usually don't want it, but newlines are different as
the line granuarity is often one at which appends are done to sections
in a "raw" fashion.  If one does not want a newline at the end, then
DELIMIT NEWLINE is still available as an alternative.