Use COLLECT-LINES / COLLECT-TEXT to simplify %rebmake.r
The make process builds a lot of command-lines from pieces, some of
which need to have their elements joined tightly (e.g. UNSPACED)
and others loosely (e.g. SPACED).

While Ren-C's concept of "opting out" with nulls and blanks made it
somewhat easier for these lines to be built, much of the rebmake code
was forced into a pattern of one giant SPACED or UNSPACED statement.
Though much better than REJOIN, it was still often a bit circuitous
to make sure segments of the code didn't add data it shouldn't.  Also,
spotting the exact points that were adding material could be hard.

This pattern in the past has been addressed for accumulating blocks
with COLLECT, where one need only look for the KEEPs inside of a
region of code.  This makes rebmake lean on the two new textual
collecting primitives COLLECT-TEXT and COLLECT-LINES to shorten and
clarify what's going on.

It also takes advantage of the general ability of enumerators like
FOR-EACH to treat blank input series as a no-op, vs testing for
that condition too often.