More "ornery" VOID! behavior, MATCH voidification
This makes voids create errors in more situations, notably requiring a
GET-WORD! or GET-PATH! to access instead of just plain word...as well
as making them require /ONLY to be used with APPEND, INSERT, MODIFY.

A previous situation in which MATCH would manage to find a match to
a rule but for a falsey input (blank, false) would have it return BAR!
in order to still be truthy in conditional slots--while giving a
"weird value to draw attention".  Given the existence of EITHER-TEST
(the routine upon which MATCH is based) it is better to return a void
and error in the moment, guiding people to use a construct if they
really need it (or just use VALUE? MATCH or NOT NULL? MATCH if only
interested in the test).

This also reviews several behaviors of voiding, like ATTEMPT doing a
voidification of null block arguments to not conflate errors with a
null result, and STOP in a CYCLE not being allowed to return null.
When the user is in control of the result of the expression such as
in these cases, it seems better to let them conflate NULL (with an
error or break) if they want to, as it is unlikely to be an accident.

Also changes TRY and OPT to not accept VOID!, with an interesting
change to make OPT on a NULL become a void, instead of just passing
through a null.  One advantage of this is to locate superfluous OPTs
(which were found by doing so).  Another is to give a tool for getting
a non-benign value (e.g. not a BLANK! from a try) in situations like a
null being reduced in a block, while not being forced to error at that
*exact* moment.  The pattern of OPT TRY can be used if one wishes nulls
to become blanks and then nulls again, matching the old behavior.