Add COLLECT-BLOCK, SPACED-TEXT, UNSPACED-TEXT
New protocols that are geared for both flexibility and performance
leverage NULL for returning a conditionally falsey "no result", with
that being transformed into blank for causing operations to be a no-op
and return NULL.

For these protocols to work, core operations like COLLECT or UNSPACED
need to return null when they have no output, instead of empty blocks
or empty strings.

This can cause some friction, especially in dealing with APIs that do
not honor the "blank-in, null-out" convention.  Some other COLLECTs
and UNSPACEDs /etc. are building results with the intention of
modifying them afterward, where special-casing the empty case to get
a empty block or text makes the calling messy.

This introduces COLLECT-BLOCK as a form of COLLECT that will return
an empty block on no KEEPs:

    >> collect-block []
    == []

    >> collect []
    // null

This replaces more complex idioms for dealing with the situation.

It also tweaks COLLECT-TEXT to obey this convention as well, assuming
you always want an empty line of text, for consistency in the naming
convention.

Adds SPACED-TEXT and UNSPACED-TEXT to try them out.  But the general
rule should be to not use these if possible, as it's more efficient
to avoid series creation and processing when not needed...and there
will be optimizations to take advantage of opting out of function
calls with BLANK! that will not be available for empty series.