Add `<blank>` annotation for retracting action invocation
Several functions were starting to follow the "BLANK!-in, NULL out"
convention.  This is a safer alternative to having the same value
signal an opt-out as well as being return (e.g. some had suggested
in R3-Alpha that `index-of` return NONE! when it was given a NONE!).
Because routines do not accept NULL input for most arguments, having
the asymmetry keeps a chain of operations from having one mysterious
failure propagate arbitrarily long without intervention.

This commit makes it easier to follow the convention, as well as
faster.  By annotating a parameter with `<blank>`, the evaluator will
not even try to run the function once arguments have been gathered...
and will immediately return NULL.