Reuse REBFRM structure for speedup/improving CHAIN
CHAIN would create a new subframe for each post-processing step in the
chain, parameterized by the output of the previous step.  This went
through the Apply mechanic, which is not as fast as being able to just
in-place reuse the REBFRM that was no longer needed for the previous
step in the function call.  However, this would disrupt error messages
or the debugger if the action were dropped--since the chain steps are
all anonymous.

Given modern factoring, this was easy to fix just by preserving the
label...and reuse mechanics make it so that allocated argument data
which is of sufficient size gets picked up immediately for a new
application.  Instead of creating a new subframe, the previous output
value is fed back in through the current evaluation REBFRM using the
same mechanism as enfix.

This inadvertently gives chains a new power, which comes pretty
naturally.  Instead of each step in the chain (after the first) being
restricted to an arity-1 function only, each step can take any number
of args. It's just that the previous step's output becomes its first
arg. Other args are evaluated at the callsite as normal.

    >> addmul: chain [:add | :multiply]
    >> addmul 1 3 5
    == 20

So multiply gets its first argument as 1 + 3... so 4. But it still
needs another argument, so it takes the 5. Previously this would have
been an error, because it would have treated it as (multiply 4) 5.