Put temporary hold on series during FOR-EACH/etc.
The semantics for how FOR-EACH, MAP-EACH, and EVERY would handle
modifications during running had not been clearly defined.  (However,
the REMOVE-EACH function was changed to not do the removal until the
entire process had been completed.)

This uses the SERIES_INFO_HOLD mechanic for these functions to hold the
series locked while they enumerate.  However, this requires hooking
into error trapping, because one doesn't want a fail() to leave the
series in a locked state.  So the enumeration is factored into a
helper that can be executed through the abstraction of the platform's
exception-handling.