Change bootstrap EXE to 8994d23
The previous bootstrapping executable for Ren-C was chosen almost two
years ago.  It required a fairly heavy amount of shim code in order
to run the build scripts that had become very "modernized" in their
use of new features.

While it held up admirably using the function composition operations of
its day, some of those features were relatively new at the time that
the executable was picked.  (Somewhat arbitrarily--Shixin committed
"whatever the executable was that day" when he committed rebmake.  But
while it may have been arbitrary, it turned out to be very prudent--as
those executables in the repository have been a useful anchor).

In any case, the shim code eventually ran up against bugs that would
only appear on some builds...and were not issues in modern Ren-C.  The
easiest move was to make a new snapshot.  For better or worse, the
"whatever the executble was that day" method is being used again...
motivated by getting past bugs in the old version vs. any particular
milestone of the EXE.  (Future bootstrap should be more methodical.)

This commits 4 new executables for Linux, OS X, Windows, and Android to
the %make/ subdirectory.