Discern empty text from blank/null in DELIMIT
The original implementation of delimit would shortcut empty text and
treat it the same as a null.  This is not always desirable...for
instance when delimiting a block of strings where each string is a
line, then an empty string represents a line of its own and will need
a newline after it.

This forces the distinction:

    ("A" = delimit ":" [_ "A" null])
    (":A:" = delimit ":" ["" "A" ""])