Error on empty protected BLOCK! appends, but not on NULL
   >> data: [a b c]
   >> protect data

   >> append data []
   ** Script error: protected value or series - cannot modify

   >> append data if false [<skip>]
   == [a b c]

Note that NULLs are non-values, and are the return result from
conditionals that do not run their branches. It is impossible to put
them in blocks, so routines like APPEND treat them as very purposeful
forms of "opting out".

Choosing to error on the empty block and not the null is based on
raising the error based on intent vs. effect. A NULL can only mean
null, it is not "incidentally null" the way a block might be
"incidentally empty".