Replace CONCOCT/composeII with new COMPOSE feature
This is a preliminary implementation of a feature that treats doubled
GROUP!s in COMPOSE as being handled with COMPOSE/ONLY semantics, while
splicing COMPOSE.

    >> block: [a b c]
    >> compose [plain: (block) only: ((block))]
    == [plain: a b c only: [a b c]]

This is supplemented by a preliminary feature for describing an escape
sequence that indicates the COMPOSE you want to do.  Only BAR! is
allowed at the moment, since using another type would currently
interfere with the evaluation (there is no EVALUATE/PART which would
be able to stop before the tail delimiter).  The BAR! has the right
properties of passing through the value and not interacting.

    >> compose '| [
        (left alone)
        ((| reverse copy [a b c] |))
        (| reverse copy [a b c] |)
        ((left alone))
    ]
    == [
        (left alone)
        [c b a]
        c b a
        ((left alone))
    ]

Future variations of this idea will allow arbitrary symbols or tags,
but will require some fundamental new evaluator functionality to do.