Introduce some usages of the double-group COMPOSE
Replaces some COMPOSE/ONLY with double-grouped forms, only getting the
/ONLY behavior on the items that need it.  Performance-wise, this is
likely faster than storing and doing the path processing.

A 2-element PATH! requires an allocation of a series node, as well as
an array large enough to store 3 cells (COMPOSE and ONLY and an end
marker).  A GROUP! inside of a GROUP! can be done using the small
series optimization...hence the overhead is one series node per
double group used.  And being able to dispatch off of a WORD! directly
is significantly faster than a PATH!.  The actual detection of the
doubled group is trivially fast compared to the cycles needed to
verify a path usage, so it works out as an improvement of both space
and performance.