Make doubling of GROUP! in PARSE compose vs. discard
When a single GROUP! is visited in PARSE during a matching rule, the
code inside the group is executed and the value is typically discarded.

This adds a feature where doubled groups, e.g. ((...)), are run and
then used as a rule.  This is very much like COMPOSE/ONLY, except with
the difference that the composition is run each time the rule is
encountered.

The basic idea of the feature is implemented by getting the GROUP!
handling code in parse to pass through a common routine.  The PARSE
code is very old and "organic", so how to do this more properly should
be investigated during ongoing cleanup efforts as time permits.