Add SET and GET /HARD for DEFAULT w/groups in path
The DEFAULT operation had a problem in working with paths that had
groups in them, because it would have to first GET the path (to see
if the value was set or not) and then SET the path.  If GET or SET
allowed evaluations, that would evaluate groups in the path twice.
For this reason, GET and SET do not allow groups in paths--it catches
what would be bugs a lot of the time.

However, it's possible to COMPOSE the path and then do a form of
setting or getting that doesn't evaluate groups.  This new refinement
is added here, called /HARD.  By doing a compose step on paths and
then doing GET/HARD and SET/HARD, evaluations happen only once.

    counter: 0
    obj: make object! [x: _]
    obj/(counter: counter + 1 'x): default [<thing>]
    did all [
       obj/x = <thing>
       counter = 1
   ]