LOGIC! false acts as PARSE FAIL, eliminate IF
This introduces a new idea that LOGIC! true is a no-op in parse (like
a BLANK!), while false causes a match to FAIL.

This is a much more flexible concept than the previous IF conditional.
IF wasn't used very often, and there there was really only one real
usage of it in Ren-C when parsing function specs.  It essentially had
to emulate EITHER with:

    if (var) [...] | if (not var) [...]

This is clearer as:

    ((either var [ [...] ][ [...] ]))

The case doesn't require LOGIC! handling to stop the rule, since there
is a rule in both branches.  But more generally, LOGIC!'s handling in
PARSE provides answers to what an IF in isolation might have been
used for in other cases.  For instance, some checking of 32-bit vs
64-bit in ENCAP did:

    if (bits = 32) [...]
        |
    if (bits = 64) [...]

Now this can be done as

    ((bits = 32)) [...]
        |
    ((bits = 64)) [...]

But it could also move the rules into the group with a switch, and
this could also make it clearer to show errors/etc.

    ((switch bits [
        32 [ [...] ]
        64 [ [...] ]
        fail "Invalid bits count"
    ]))