Make COMPOSE/DEEP not copy levels that don't compose
COMPOSE/DEEP would previously make copies of all levels in the source
that were not insertions, even though it didn't make copies of the
insertions.  (This meant that subtrees which had no GROUP!s anywhere
in them would be duplicated, even though the overall operation did
not have COPY/DEEP semantics.)

Typically this is not necessary.  This makes an experimental change
(which would be fairly easy to back out) that abandons copies when
nothing underneath that level needs a copy.  The topmost level is
excluded--it is copied even if there are no groups in it.