Finish reclaiming of VALUE_FLAG_THROWN
VALUE_FLAG_THROWN is a scarce header bit, being effectively wasted on
something that should only be on one value in the system at a time.
However, at the time of its introduction it seemed to make sense--as
there was no way to mark an arbitrary value as thrown otherwise, in
order to generate asserts if it was accessed by code that thought it
was just a normal value.  A bit was the only way to inject checking.

This commit applies an alternative strategy by having the debug build
sometimes move thrown values to a global location other than the cell
output.  That's inefficient in the general case, but by doing it half
the time in the debug build and putting garbage in the output slot,
it's possible to find most violations...without having to have special
checks for it (the usual checks for garbage suffice).

So this builds on prior work to actually reclaim VALUE_FLAG_THROWN
completely, with a more robust solution to the problems that the bit
was originally trying to solve.