Move holding flag into frame source, not frame
The same array can appear in multiple recursions of the evaluator.  So
if a higher level has already put a temporary hold on it against
modification, a lower level shouldn't be involved as well, because it
would release the bit prematurely.

An attempt to solve this was made by having a frame flag which would
indicate if a frame had taken a hold.  This was based on the idea that
there would be a single array which would be held for the entire
duration of a frame.  This concept changed, where it might be possible
to inject multiple arrays and have each one held for as long as it
was running.  And some of the invariants weren't being preserved
properly, because multiple frame recursions could use the same "source"
information (va_list, arrays).

It led to some stray holds being kept, which were not noticed due to
source immutability.  Adding a more conservative `const` strategy
exposed these bugs, so the flag had to be moved in to the source vs.
being a bit on the frame itself.