Simplify and expand STATIC_ASSERTs
There was an attempt to have a static assert functionality that would
work in C, but it would only work in functions.  Because any C++ build
can check the condition--and they're internal conditions to the
evaluator codebase--it seems better to use the more powerful C++ static
assert and just let it be a no-op in the C build.

So this replaces some verbal comments on which flags had to have the
same values as others with asserts right at the flag.  Also has some
random comment fixes, as well as the introduction of VALUE_FLAG_CONST
(while not involving the support code for the feature yet, there was
just rearrangement of flags involving it that was clearer to do in
this commit).