CONST and MUTABLE
This introduces a new kind of read-only-ness to the system.  Unlike
PROTECT which affects a series through all values which might access
it, CONST access is on a per-value basis.

    >> data: mutable [a b c]
    == [a b c]

    >> data-readonly: const data
    == [a b c]

    >> append data-readonly 'd
    ** Access Error: value is CONST (see MUTABLE): [a b c]

    >> append data 'd
    == [a b c d]

    >> data-readonly
    == [a b c d]

The implementation is preliminary and experimental, to get experience
with if this works in the field.  As it is a safety feature only, it
is being committed in a somewhat non-rigorous state.  More rigor
would likely demand bigger changes, that would need to be informed by
spending some time using it.