Add C-DEBUG-TICK native, first phase Travis https fetch
This allows you to get the current tick count in a debug build, which
could be used to guide things like when to choose to do usermode debug
spew (if a crashing tick were known, it could narrow the range of when
output might be relevant).

Also, on platforms that build Rebol twice, this does a smoke test of
reading the https example.com page before running the second build, to
notice bad builds earlier (and with address sanitizer still running).