Add ENBLOCK, ENGROUP, BLOCKIFY, GROUPIFY
This adds four optimized natives for the fairly common task of making
groups and blocks out of values.  ENBLOCK will enclose any value in a
block, regardless of whether it initially was a block or not--and
ENGROUP will do the same for a group.  While BLOCKIFY will pass back a
block if given one, otherwise enclose it in a block...same for GROUPIFY
but it will be a no-op on GROUP!.

Any of these routines treat null input as a case to give back an empty
block (blank input will put the blank in a block, so this could be
important for BLOCKIFY OPT ... scenarios, where a blank indicates an
empty block is desired).

These routines pick up some slack from the fact that TO BLOCK! and
TO GROUP! are kind of up in the air definitionally.  These words aren't
going to be taken for anything else, so they don't do any real harm
being used for this.

(The only real question would be if BLOCKIFY and GROUPIFY should copy
their input if it's a block, to give back the invariant that the series
given back is "new".  But for those who need to do it, copying singular
arrays is relatively inexpensive.)