Faster rebArgR() primitive for rebR(rebArg())
After debating whether the API should receive non-API handles that
would point directly at argument slots in the frame, the resolution
was that it would not provide enough flexibility (e.g. if the user
wanted to change the lifetime of the handle to be indefinite--which
can't be done with a pointer into a frame's actual arg memory).

However, an optimization can be done by returning the argument as
an opaque const void* designed for use as a variadic argument.
This way the user never saves the value into a REBVAL*, but still
does not have to worry about freeing it.  This is good for if there
is just one question of the argument, such as "what is your unboxed
integer value".

Also adds test for closed Ren-C issue 633.