Make blocks in %sysobj.r object MUTABLE
Some static blocks were being used in %sysobj.r, and while two were
converted to use COPY in the early moments of the CONST/MUTABLE build,
this overlooked a couple of others.  It is now possible to use MUTABLE
instead, even in the bootstrap build...as it is a no-op in the shim.

Also fixes a problem with reporting errors during startup with the
"die" function.