Efficient LITERAL!, up to 3 quote levels w/no new cell
Value cells reserved a full byte for the "cell kind", despite there
being fewer than 64 unique base system types.  This takes advantage of
the extra space in the byte by adding 64 to it, in order to make it
a quoted version of the type.  This can be done 3 times, in order to
get 3 levels of escaping represented in a single cell.  After that
point, a separate payload must be allocated and the cell split into
two parts, to represent the higher escape levels.

Implementing this safely takes advantage of the ability to build as
C++, by creating a new datatype through which the VAL_TYPE() operation
cannot be used, only the CELL_KIND() operation.  This is because the
VAL_TYPE() will report REB_LITERAL for any kind bytes that are above
64, but certain code (like molding or comparison) wants to be able
to work with the payload of the underlying type of a literal--after it
has already understood that it is a literal.