Streamline hook naming in %types.r
R3-Alpha would automatically generate tables of hook functions for
various behaviors of datatypes.  These tables sometimes would delegate
responsibility for something like path dispatch to a function that was
specific to the datatype itself, and other times to a common function
for a class (e.g. PICK was implemented the same for GROUP! and BLOCK!)

This simplifies and documents how those hooks work, and how the types
table generates the names it does.