Avoid mod in %c-eval.c for type bytes, simplify rebUneval()
Because values at quoting levels of 1-3 use higher values of the low
level kind byte in a cell, they can't be confused with the bytes of
an unquoted value.  The evaluator loop only cares about behavioral
interaction with the low 64 kind values, as all literals have the
same behavior (remove one level of the literalness).  Since the types
are being put through a switch that can have a default, there is no
reason to use VAL_TYPE_RAW() which clips the types to REB_LITERAL
if they are > 64... it can just use the KIND_BYTE() directly and then
let all higher byte values hit the default case in the switch.

The rebUneval() routine had previously had to construct GROUP!s to get
arbitrary values into a single cell slot.  Now it can more optimally
just have a version of the value with an escape level; this will fit
into a single cell almost all of the time.

Also tweaks LIT to have the meaning of what QUOTE historically did
instead of UNEVAL.  This opens the possibility that QUOTE could be
changed to the more useful "add a level of quote to an evaluated
argument" function, based on the idea of changing LITERAL! to QUOTED!
in order to use more correct terminology.