Unify apostrophe escaping with generalized escaping
This takes the technique used by the scanner to integrate with the
generalized escaping mechanism and allows one to use it with apostrophe
instead of backslash.

Backslash was the character chosen first when implementing generalized
escaping, but it is potentially subject to change.

The only types which have apostrophe-flavored cell representations are
LIT-BAR!, LIT-WORD!, and LIT-PATH!...in accordance with history.
However, representations which cannot match that fall back upon generic
escaping if necessary:

    >> 'x
    == x

    >> ''x
    == \x

    >> first ['x]
    == 'x

    >> first [''x]
    == \\x

Even though LIT-SET-WORD! and LIT-GROUP! etc. do not exist to be
manifests for the apostrophe forms, usages which aren't storing the
intermediate, and merely using it to avoid evaluation for one step,
won't have problems with it:

    >> block: copy []
    >> append/only block '(1 + 2)
    == [(1 + 2)]

This means that--for now--apostrophe can be used to quote other forms
of previously unquotable items:

    >> a: 10

    >> foo: function [] [set 'a: 20 | print ["a is" a]]

    >> foo
    a is 20

    >> print ["a is" a]
    a is 10

Previously such stunts would have to be accomplished via QUOTE (which
is now being rethought as LIT)

    >> lit 'a
    == 'a

    >> lit ''''(1 + 2)
    == \\\\(1 + 2)

The choice of escaping character is an open issue and being discussed.